# Restore energy loss in indirect (IBL) specular lighting via multiscattering approximation