Currently, for metal surfaces (MeshPhysicsMaterial, MeshStandardMaterial), as the roughness of the material increases, the surface darkens if lit via IBL+PMREM (possibly #7489?). For fully metal surfaces (conductors), all incoming irradiance should equal outgoing radiance (F0=1) -- like a mirror (metalness=1, roughness=0). Even if the metal is rough (metalness=1, roughness=1), the same amount of energy should be reflected off of the material.

There is the [white furnace test](https://google.github.io/filament/Filament.md.html#toc4.7.2), which has a uniformly lit enironment with fully metal, white spheres with varying degrees of roughness. As these surfaces are fully metal, all incoming light must be reflected, and should be indistinguishable from the environment. If the spheres are visible, then some energy is either lost or gained.

Currently, energy is lost as metal surfaces increase in roughness:

![furnace-before](https://user-images.githubusercontent.com/641267/51778698-0e730100-20b8-11e9-8587-e06f15ce054c.png)

BRDFs only account for a single scattering of light -- for rough, metal surfaces, light will bounce around microfacets after the initial scattering before reaching the viewer, which isn't accounted for in BRDF models.

![diagram_single_vs_multi_scatter](https://user-images.githubusercontent.com/641267/51778817-9eb14600-20b8-11e9-8c5e-fa662c80f161.png)
via [Filament's beautiful docs](https://google.github.io/filament/Filament.md.html)


There are several sampling and preprocessing methods that aid in real-time approximation of multiscattering light, but ultimately, I was able to get Fdez-Agüera's method ([A Multiple-Scattering Microfacet
Model for Real-Time Image-based Lighting](http://jcgt.org/published/0008/01/03/)) working well, and used a [DFG approximation from Filament](https://google.github.io/filament/Filament.md.html#lighting/imagebasedlights/distantlightprobes) instead of a DFG LUT .

![comparison-beforeaftersmall](https://user-images.githubusercontent.com/641267/51778960-5d6d6600-20b9-11e9-9311-ca8b060dc8fe.png)

Above, on the left is r100, on the right is this branch -- from top to bottom, metal surface with increasing roughness, rough surface with increasing metalness, metal surface with increasing roughness, nonmetal increasing rough surface). Much more energy is retained on rough, metal objects with this change.


![specular-multiscatter-pieces-small](https://user-images.githubusercontent.com/641267/51779069-f7cda980-20b9-11e9-81a1-8be4f45c07a6.png)

Above is the single-scattering specular indirect lighting only (left), only the multiscattering (center), and both (combined). Currently in r100, only the single-scattering is visible.

![furnace-after](https://user-images.githubusercontent.com/641267/51779164-93f7b080-20ba-11e9-99f3-d765cad11f4f.png)

Perhaps the least exciting image, above is the furnace test with these changes. All incoming light equals the outgoing light, making the spheres invisible.

Now a handful of questions:

* Should this be MeshPhysicalMaterial only?
* What should the conditions for this be? Currently, it's `defined( USE_ENVMAP ) && defined( ENVMAP_TYPE_CUBE_UV )`, but mostly I'm checking to see if a radiance map exists.
* Needs a good correctness check
* Performance check (seems good, but needs a rigorous test)
* Is the DFG approximation valid?
* Do we need to separate the indirect specular and diffuse lighting functions? They're combined here to save computations.
* Maybe this should only be on desktop, depending on perf