@mrdoob 2019-01-26
@bhouston @spidersharma03 do these changes look good to you?