@jsantell 2019-01-26
Some perf testing, it looks like this might be slighty slightly less performant than the current implementation, but practically the same. Haven't given too much of an eye to optimization yet. It'd be helpful if someone could check it out on iOS.

Test case, 64 objects mostly filling the screen

* https://indirect-perf.glitch.me/ (for testing PR)
* https://indirect-perf.glitch.me/?r100 (for testing current r100 build)

Mobile Results

device | browser | current FPS | proposed FPS |
----- | --- | --- | --- |
Pixel 2 | Chrome r71 | ~59 | ~59
Pixel 2 | Firefox v64 | ~57 | ~57
Pixel 1 | Chrome r71 | ~39 | ~38
Pixel 1 | Firefox v64 | ~47 | ~46

Linux Laptop, 800x800

browser | DPR | current FPS | proposed FPS |
----- | --- | --- | --- |
Firefox v64 | 1 | ~58 | ~58
Firefox v64 | 2 |  ~32 | ~30
Firefox v64 | 3 |  ~19 | ~19
Chromium v71 | 2 | ~44 | ~43

> If you want to refactor the shader code, that would be fine with me. I find the `#define RE_Direct` chunks, etc., somewhat confusing.

Sounds great, I can take care in a follow up