@donmccurdy 2019-01-26
iPhone SE, Safari: 33fps before and after.