@bhouston 2019-01-26
Is there any way this could be precalculated into the ibl map, even in an
approximate fashion just to maintain energy?

On Sat, Jan 26, 2019, 3:54 PM Don McCurdy <notifications@github.com wrote:

> iPhone SE, Safari: 33fps before and after.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/mrdoob/three.js/pull/15644#issuecomment-457865173>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAj6_Y5Ow25EnqG9rE62eM14HNvsq4Glks5vHMCcgaJpZM4aT26j>
> .
>