@jsantell 2019-01-28
Updated PR with `preserveEnergy` boolean on MeshPhysicalMaterial that `#define PRESERVE_ENERGY`'s when set with a PMREM so it could possibly be more opt in. Also consolidated the integrated BRDF generation code which was mostly what the current indirect specular was generating.

Also as a quick double check, this is what the DFG approximation looks like generated in `integrateBRDF`:

![localhost_8080_examples_texture html 5](https://user-images.githubusercontent.com/641267/51819659-0964ac00-2288-11e9-9de8-794289ff5f5f.png)