@WestLangley 2019-01-28
>Shouldn't preserveEnergy be true by default? The improved output seems worth the slight loss of fps.

I'd hardwire it, then, and remove the `preserveEnergy` flag.