@jsantell 2019-01-28
> Shouldn't `preserveEnergy` be `true` by default? The improved output seems worth the slight loss of fps.

Wasn't sure about the (small) performance delta, and figured this would be the best way for opt in -- happy to remove the flag if should be enabled for all MeshPhysicalMaterials with PMREMs -- which maybe MeshPhysicalMaterial over MeshStandardMaterial is sufficient opt-in for potentially higher cost for more accurate renderings