@jsantell 2019-01-28
> I think this should apply to both `MeshStandardMaterial` and `MeshPhysicalMaterial` and be enabled by default. Good looking by default is best. If people need extra ms they could opt-out.

SGTM -- should it continue being a boolean for folks to disable?