@mrdoob 2019-01-28
Is there any other reason for that boolean apart from a few extra ms? If not, I would remove the boolean altogether (and avoid complexity).