@jsantell 2019-01-29
Updated the styling of the glsl strings :+1: