@mrdoob 2019-01-29
The `webgl_materials_standard` looks definitely much better.

Before:
<img width="910" alt="screen shot 2019-01-29 at 9 17 43 pm" src="https://user-images.githubusercontent.com/97088/51941324-3a6de980-240c-11e9-9e31-68f3a9216afb.png">
https://raw.githack.com/mrdoob/three.js/r100/examples/webgl_materials_standard.html

After:
<img width="910" alt="screen shot 2019-01-29 at 9 16 48 pm" src="https://user-images.githubusercontent.com/97088/51941379-58d3e500-240c-11e9-84da-25dc7a31fd76.png">
https://raw.githack.com/mrdoob/three.js/dev/examples/webgl_materials_standard.html

However, `webgl_materials_envmaps_exr` and `webgl_materials_envmaps_hdr` now have some sort of dark fresnel?

Before:
<img width="909" alt="screen shot 2019-01-29 at 9 30 21 pm" src="https://user-images.githubusercontent.com/97088/51941777-53c36580-240d-11e9-9bdd-292c64f56422.png">
https://raw.githack.com/mrdoob/three.js/r100/examples/webgl_materials_envmaps_exr

After:
<img width="910" alt="screen shot 2019-01-29 at 9 30 04 pm" src="https://user-images.githubusercontent.com/97088/51941875-88cfb800-240d-11e9-98ae-102a27b2dfdc.png">
https://raw.githack.com/mrdoob/three.js/dev/examples/webgl_materials_envmaps_exr

Before:
<img width="910" alt="screen shot 2019-01-29 at 9 30 46 pm" src="https://user-images.githubusercontent.com/97088/51941925-aef55800-240d-11e9-8d36-7e1132b46552.png">
https://raw.githack.com/mrdoob/three.js/r100/examples/webgl_materials_envmaps_hdr

After:
<img width="910" alt="screen shot 2019-01-29 at 9 31 05 pm" src="https://user-images.githubusercontent.com/97088/51941946-be74a100-240d-11e9-8fd8-22532a144ff5.png">
https://raw.githack.com/mrdoob/three.js/dev/examples/webgl_materials_envmaps_hdr