@jsantell 2019-01-29
@mrdoob noted, will take a look!