@WestLangley 2019-01-29
>`webgl_materials_standard` looks definitely much better.

What exactly are you seeing? I see no visible difference. Maybe I am missing something...

I have only seen significant visible difference in real-world examples (that is, non-furnace tests) for rough-metallic surfaces. Maybe @jsantell has seen otherwise...