@jsantell 2019-01-30
> > `webgl_materials_standard` looks definitely much better.
> 
> What exactly are you seeing? I see no visible difference. Maybe I am missing something...
> 
> I have only seen significant visible difference in real-world examples (that is, non-furnace tests) for rough-metallic surfaces. Maybe @jsantell has seen otherwise...

The only lighting that's changed is an extra specular lobe for metal surfaces so the change is most noticeable on very rough metallic surfaces, where previously the specular light was lost