@jsantell 2019-01-30
Potential fix for the dark fresnel ring: https://github.com/mrdoob/three.js/pull/15673