@Mugen87 2019-02-01
@jsantell I think we should note in the [migration guide](https://github.com/mrdoob/three.js/wiki/Migration-Guide) that the rendering of `MeshPhysicsMaterial`/ `MeshStandardMaterial` slightly changed. How would you phrase the most important aspects of the PR?