@Mugen87 2019-02-01
Perfect! I've added your suggestion to the guide.